pyproject.toml
[project]
name = "court-jester-selector"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.2"
dependencies = [
    "alembic==1.15.2",
    "python-telegram-bot[all]==22.0",
    "sqlmodel==0.0.24",
]

[dependency-groups]
dev = [
    "asyncpg==0.30",
]
